Fixed loading non-exists files

+++ app/src/main/java/ru/magflayer/spectrum/domain/interactor/ColorPhotoInteractor.java
@@ -1,5 +1,7 @@
 package ru.magflayer.spectrum.domain.interactor;
 
+import java.io.File;
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.inject.Inject;
@@ -22,11 +24,25 @@ public class ColorPhotoInteractor {
     }
 
     public Observable<List<ColorPhotoEntity>> loadColorPhotos() {
-        return photoRepository.loadPhotos();
+        return photoRepository.loadPhotos()
+                .map(entities -> {
+                    List<ColorPhotoEntity> photoEntities = new ArrayList<>();
+
+                    for (ColorPhotoEntity entity : entities) {
+                        if (new File(entity.getFilePath()).exists()) {
+                            photoEntities.add(entity);
+                        } else {
+                            photoRepository.removePhoto(entity);
+                        }
+                    }
+
+                    return photoEntities;
+                });
     }
 
     public Observable<Boolean> removeColorPhoto(final ColorPhotoEntity entity) {
-        return photoRepository.removePhoto(entity);
+        return photoRepository.removePhoto(entity)
+                .map(success -> new File(entity.getFilePath()).delete());
     }
 
     public Observable<ColorPhotoEntity> loadColorPhoto(final String filePath) {
